# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Nöbetçi Eczane API
  version: 0.1.0
  description: |
    Nöbetçi eczane bulma uygulamasının API sözleşmesi.

    **Bu dosya tek doğru kaynaktır.** Frontend tipleri buradan üretilir:
    `npm run gen:types` → `src/shared/api-types.d.ts`

    Elle tip yazılmaz. Önce bu dosya değişir, sonra tipler üretilir, sonra kod yazılır.

    Veri kaynağı ve tazeleme davranışı için: `docs/adr/001-veri-kaynagi.md`,
    `docs/adr/003-veri-tazeleme.md`
servers:
  - url: http://localhost:8787
    description: Yerel geliştirme (wrangler dev)
  - url: https://nobetci-eczane.workers.dev
    description: Production
...
paths:
...
  /api/admin/session:
    post:
      tags: [admin]
      operationId: adminLogin
      summary: Yönetici girişi
      description: |
        Başarılı girişte `HttpOnly; Secure; SameSite=Strict` oturum çerezi kurulur.
        Kimlik bilgileri kodda değil, ortam değişkenlerinden gelir; parola PBKDF2 ile
        hash'lenmiş saklanır. Karşılaştırma sabit zamanlıdır.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              additionalProperties: false
              properties:
                username: { type: string, minLength: 1, maxLength: 64 }
                password: { type: string, minLength: 1, maxLength: 256 }
      responses:
        "204":
          description: Giriş başarılı, oturum çerezi kuruldu
          headers:
            Set-Cookie:
              schema: { type: string }
              example: "session=…; HttpOnly; Secure; SameSite=Strict; Path=/; Max-Age=28800"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401":
          description: Kimlik bilgileri hatalı
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
    delete:
      tags: [admin]
      operationId: adminLogout
      summary: Oturumu kapat
      security: [{ sessionCookie: [] }]
      responses:
        "204": { description: Oturum kapatıldı }
...
  /api/admin/known-coords:
    get:
      tags: [admin]
      operationId: listKnownCoords
      summary: Koordinatı zaten bilinen eczane slug'ları
      description: |
        Çekim komutu bunu kullanarak **gereksiz harita isteği yapmaz** (ADR-005).
        Verilen slug'lardan hangilerinin lat/lng'si dolu, onu döner.
      security: [{ sessionCookie: [] }]
      parameters:
        - name: city
          in: query
          required: true
          schema: { type: integer, minimum: 1, maximum: 81 }
        - name: slugs
          in: query
          required: true
          description: Virgülle ayrılmış slug listesi.
          schema: { type: string, maxLength: 20000 }
      responses:
        "200":
          description: Koordinatı bilinen slug'lar
          content:
            application/json:
              schema:
                type: object
                required: [slugs]
                additionalProperties: false
                properties:
                  slugs:
                    type: array
                    items: { type: string }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/admin/import:
    post:
      tags: [admin]
      operationId: importDuties
      summary: Çekilen nöbet listesini topluca yaz
      description: |
        Tek çağrıda: bilinmeyen ilçeleri ekler, eczaneleri slug'a göre upsert eder,
        o güne ait nöbetleri yazar.

        **Elle girilmiş koordinat korunur:** `coordSource = 'manual'` olan eczanenin
        lat/lng'si bu uçtan ASLA ezilmez (ADR-005).

        O gün için daha önce yazılmış `source='edevlet'` nöbetleri silinip yeniden
        yazılır — böylece kaynaktan çıkarılan bir eczane bizde asılı kalmaz.
        `source='manual'` nöbetlere dokunulmaz.
      security: [{ sessionCookie: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ImportRequest" }
      responses:
        "200":
          description: Yazıldı
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ImportResult" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404":
          description: İl desteklenmiyor
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /api/admin/scrape/trigger:
    post:
      tags: [admin]
      operationId: triggerScrape
      summary: e-Devlet çekimini elle tetikle
      description: |
        Çekim Cloudflare'in **dışında** koşar (GitHub Actions). Bu uç yalnızca
        `repository_dispatch` çağrısı yapar — tek subrequest, HTML ayrıştırma yok.
        Gerekçe: ücretsiz planda invocation başına 10 ms CPU / 50 dış subrequest
        sınırı 81 ilin ayrıştırmasına yetmez (ADR-006).

        Kaynağa saygı gereği **10 dakikada bir** tetiklenebilir; daha sık denemede
        `429` döner. Sonuç anında gelmez: iş akışı birkaç dakika sonra
        `POST /api/admin/scrape/runs` ile koşu kayıtlarını yazar.
      security: [{ sessionCookie: [] }]
      requestBody:
        required: true
        description: Alanların hepsi varsayılanlı; boş gövde `{}` gönderilebilir.
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ScrapeTriggerInput" }
      responses:
        "202":
          description: İş akışı tetiklendi
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ScrapeTriggerResult" }
              example:
                dispatchedAt: "2026-08-08T11:12:03Z"
                scope: "tum"
                days: "ikisi"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
        "502":
          description: GitHub'a ulaşılamadı veya dispatch reddedildi
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "503":
          description: Tetikleme yapılandırılmamış (`GITHUB_DISPATCH_TOKEN` / `GITHUB_REPO` yok)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /api/admin/scrape/runs:
    get:
      tags: [admin]
      operationId: listScrapeRuns
      summary: Son çekim koşuları
      description: |
        Panelin çekim sekmesi bunu gösterir: hangi il, hangi gün, ne sonuç verdi.
        En yeni koşu başta döner.
      security: [{ sessionCookie: [] }]
      parameters:
        - name: city
          in: query
          description: Plaka kodu. Verilmezse tüm iller.
          schema: { type: integer, minimum: 1, maximum: 81 }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
      responses:
        "200":
          description: Koşu listesi
          content:
            application/json:
              schema:
                type: object
                required: [items]
                additionalProperties: false
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/ScrapeRun" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      tags: [admin]
      operationId: recordScrapeRun
      summary: Bir çekim koşusunun sonucunu yaz
      description: |
        Çekim komutu her il-gün bitiminde bunu çağırır. Kayıt **tamamlanmış** koşuyu
        anlatır; `outcome` alanı ne olduğunu söyler. Sessiz bozulma buradan görülür:
        `rowsFound` düşerse ya da `error` birikirse kaynak değişmiş demektir.
      security: [{ sessionCookie: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ScrapeRunInput" }
      responses:
        "201":
          description: Koşu kaydedildi
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ScrapeRun" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/admin/me:
    get:
      tags: [admin]
      operationId: getAdminMe
      summary: Oturum geçerli mi
      description: Panel açılışında oturumun hâlâ geçerli olduğunu doğrulamak için.
      security: [{ sessionCookie: [] }]
      responses:
        "200":
          description: Oturum geçerli
          content:
            application/json:
              schema:
                type: object
                required: [username, expiresAt]
                additionalProperties: false
                properties:
                  username: { type: string }
                  expiresAt: { type: string, format: date-time }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/admin/pharmacies:
    get:
      tags: [admin]
      operationId: listAdminPharmacies
      summary: Eczane kayıtları
      description: |
        Panelin eczane defteri. Nöbet ataması ayrı bir kayıttır (`/api/admin/duties`);
        buradaki liste eczanenin kendisidir, nöbette olup olmadığından bağımsızdır.

        Veri 81 ili kapsadığı için liste **her zaman sayfalıdır**: `total` toplam
        eşleşmeyi, `items` bu sayfayı verir.
      security: [{ sessionCookie: [] }]
      parameters:
        - name: q
          in: query
          description: Ad veya adres içinde arama.
          schema: { type: string, maxLength: 120 }
        - name: city
          in: query
          description: Plaka kodu. Verilmezse tüm iller.
          schema: { type: integer, minimum: 1, maximum: 81 }
        - name: district
          in: query
          schema: { type: string, pattern: "^[0-9]{4}$" }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 200, default: 100 }
        - name: offset
          in: query
          schema: { type: integer, minimum: 0, default: 0 }
      responses:
        "200":
          description: Eczane listesi
          content:
            application/json:
              schema:
                type: object
                required: [total, items]
                additionalProperties: false
                properties:
                  total: { type: integer }
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/AdminPharmacy" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      tags: [admin]
      operationId: createPharmacy
      summary: Eczane ekle
      security: [{ sessionCookie: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PharmacyInput" }
      responses:
        "201":
          description: Oluşturuldu
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminPharmacy" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "409":
          description: Aynı ad ve ilçede kayıt zaten var
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /api/admin/pharmacies/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: integer, minimum: 1 }
    patch:
      tags: [admin]
      operationId: updatePharmacy
      summary: Eczane güncelle
      security: [{ sessionCookie: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PharmacyInput" }
      responses:
        "200":
          description: Güncellendi
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminPharmacy" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [admin]
      operationId: deletePharmacy
      summary: Eczane sil
      description: Eczaneye bağlı nöbet kayıtları da silinir.
      security: [{ sessionCookie: [] }]
      responses:
        "204": { description: Silindi }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/admin/duties:
    get:
      tags: [admin]
      operationId: listDuties
      summary: Nöbet atamaları
      description: |
        Takvimde bir güne tıklandığında açılan liste. 81 ilde bir günün nöbeti
        binlerce satır olabildiği için sonuç **sayfalıdır** ve pratikte il
        filtresiyle kullanılır.
      security: [{ sessionCookie: [] }]
      parameters:
        - name: date
          in: query
          description: "Nöbet günü (YYYY-MM-DD, TRT). Verilmezse bugün."
          schema: { type: string, pattern: "^[0-9]{4}-[0-9]{2}-[0-9]{2}$" }
        - name: city
          in: query
          description: Plaka kodu. Verilmezse tüm iller.
          schema: { type: integer, minimum: 1, maximum: 81 }
        - name: district
          in: query
          schema: { type: string, pattern: "^[0-9]{4}$" }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 200, default: 100 }
        - name: offset
          in: query
          schema: { type: integer, minimum: 0, default: 0 }
      responses:
        "200":
          description: O güne ait nöbet atamaları
          content:
            application/json:
              schema:
                type: object
                required: [date, total, items]
                additionalProperties: false
                properties:
                  date: { type: string }
                  total: { type: integer, description: Filtreye uyan toplam kayıt. }
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/DutyShift" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      tags: [admin]
      operationId: createDuty
      summary: Nöbet ata
      description: |
        Nöbet penceresi `date` gününün 08:00 TRT'sinden ertesi gün 08:00 TRT'sine kadardır.
        Aynı eczaneye aynı gün için ikinci atama yapılamaz.
      security: [{ sessionCookie: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DutyInput" }
      responses:
        "201":
          description: Atandı
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DutyShift" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: Bu eczane o gün için zaten nöbetçi
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /api/admin/duties/calendar:
    get:
      tags: [admin]
      operationId: getDutyCalendar
      summary: Aylık nöbet takvimi
      description: |
        Panelin takvim ızgarasını besler: ayın her günü için nöbet sayısı ve
        kaynağa göre dağılım. **Eksik gün buradan görülür** — sayısı 0 olan gün
        ızgarada işaretlenir.

        Yalnızca kayıt bulunan günler döner; boş günler dizide yer almaz.
      security: [{ sessionCookie: [] }]
      parameters:
        - name: month
          in: query
          required: true
          description: "Ay (YYYY-MM, TRT takvimi)."
          schema: { type: string, pattern: "^[0-9]{4}-[0-9]{2}$" }
          example: "2026-08"
        - name: city
          in: query
          description: Plaka kodu. Verilmezse tüm iller.
          schema: { type: integer, minimum: 1, maximum: 81 }
      responses:
        "200":
          description: Gün başına nöbet sayıları
          content:
            application/json:
              schema:
                type: object
                required: [month, cityCode, days]
                additionalProperties: false
                properties:
                  month: { type: string, example: "2026-08" }
                  cityCode: { type: [integer, "null"] }
                  days:
                    type: array
                    items: { $ref: "#/components/schemas/CalendarDay" }
              example:
                month: "2026-08"
                cityCode: 7
                days:
                  - date: "2026-08-08"
                    count: 37
                    edevletCount: 37
                    manualCount: 0
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/admin/duties/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: integer, minimum: 1 }
    delete:
      tags: [admin]
      operationId: deleteDuty
      summary: Nöbet atamasını kaldır
      security: [{ sessionCookie: [] }]
      responses:
        "204": { description: Silindi }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: session
...
  responses:
    BadRequest:
      description: Geçersiz istek parametresi
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: Oturum yok veya geçersiz
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Kayıt bulunamadı
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    TooManyRequests:
      description: Çok fazla deneme
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Error:
      type: object
      required: [error]
      additionalProperties: false
      properties:
        error:
          type: object
          required: [code, message]
          additionalProperties: false
          properties:
            code:
              type: string
              enum:
                - bad_request
                - unauthorized
                - not_found
                - city_not_supported
                - rate_limited
                - source_unavailable
                - dispatch_failed
                - not_configured
                - internal
            message:
              type: string
              description: Kullanıcıya gösterilebilir Türkçe mesaj.
            details:
              type: array
              description: Alan bazlı doğrulama hataları.
              items:
                type: object
                required: [path, message]
                additionalProperties: false
                properties:
                  path: { type: string }
                  message: { type: string }
...
    ImportRequest:
      type: object
      required: [cityCode, dutyDate, source, items]
      additionalProperties: false
      properties:
        cityCode: { type: integer, minimum: 1, maximum: 81 }
        dutyDate: { type: string, pattern: "^[0-9]{4}-[0-9]{2}-[0-9]{2}$" }
        source: { type: string, enum: [edevlet] }
        items:
          type: array
          minItems: 1
          maxItems: 500
          items: { $ref: "#/components/schemas/ImportItem" }
    ImportResult:
      type: object
      required:
        [
          districtsCreated,
          pharmaciesCreated,
          pharmaciesUpdated,
          dutiesWritten,
          dutiesRemoved,
          coordsPreserved,
        ]
      additionalProperties: false
      properties:
        districtsCreated: { type: integer }
        pharmaciesCreated: { type: integer }
        pharmaciesUpdated: { type: integer }
        dutiesWritten: { type: integer }
        dutiesRemoved:
          type: integer
          description: Kaynaktan çıkarıldığı için silinen önceki edevlet nöbetleri.
        coordsPreserved:
          type: integer
          description: Elle girildiği için korunan koordinat sayısı.

    # ─── Panel şemaları (ADR-004) ──────────────────────────────────────────
...
    ScrapeTriggerInput:
      type: object
      additionalProperties: false
      properties:
        scope:
          type: string
          description: '`tum` → 81 il. Tek il için plaka kodunu metin olarak ver (`"7"`).'
          pattern: "^(tum|[1-9][0-9]?)$"
          default: "tum"
        days:
          type: string
          enum: [bugun, yarin, ikisi]
          default: ikisi
    ScrapeTriggerResult:
      type: object
      required: [dispatchedAt, scope, days]
      additionalProperties: false
      properties:
        dispatchedAt: { type: string, format: date-time }
        scope: { type: string }
        days: { type: string, enum: [bugun, yarin, ikisi] }
    ScrapeRunInput:
      type: object
      required: [cityCode, dutyDate, startedAt, outcome]
      additionalProperties: false
      properties:
        cityCode: { type: integer, minimum: 1, maximum: 81 }
        dutyDate: { type: string, pattern: "^[0-9]{4}-[0-9]{2}-[0-9]{2}$" }
        startedAt:
          type: string
          format: date-time
          description: "Milisaniyesiz UTC: `YYYY-MM-DDTHH:MM:SSZ` (CLAUDE.md kota kuralı)."
        outcome: { type: string, enum: [ok, partial, error] }
        rowsFound: { type: integer, minimum: 0, default: 0 }
        pharmaciesNew: { type: integer, minimum: 0, default: 0 }
        dutiesWritten: { type: integer, minimum: 0, default: 0 }
        coordsFetched:
          type: integer
          minimum: 0
          default: 0
          description: Bu koşuda kaynağa atılan harita isteği sayısı.
        rowsSkipped: { type: integer, minimum: 0, default: 0 }
        errorMessage: { type: [string, "null"], maxLength: 500 }
    ScrapeRun:
      type: object
      required:
        [
          id,
          cityCode,
          cityName,
          dutyDate,
          startedAt,
          finishedAt,
          outcome,
          rowsFound,
          pharmaciesNew,
          dutiesWritten,
          coordsFetched,
          rowsSkipped,
          errorMessage,
        ]
      additionalProperties: false
      properties:
        id: { type: integer }
        cityCode: { type: integer }
        cityName: { type: string, example: "Antalya" }
        dutyDate: { type: string, example: "2026-08-08" }
        startedAt: { type: string, format: date-time }
        finishedAt: { type: [string, "null"], format: date-time }
        outcome: { type: string, enum: [ok, partial, error] }
        rowsFound: { type: integer }
        pharmaciesNew: { type: integer }
        dutiesWritten: { type: integer }
        coordsFetched: { type: integer }
        rowsSkipped: { type: integer }
        errorMessage: { type: [string, "null"] }
...